use whatlang with a 0.93 confidence to discard non-English paragraphs
